+++ noodles-cram/src/codecs/rans_4x8/order.rs
@@ -11,9 +11,12 @@ impl fmt::Display for TryFromByteError {
 
 impl error::Error for TryFromByteError {}
 
+/// The number of bytes of context used to compute frequencies.
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 pub enum Order {
+    /// Order-0.
     Zero,
+    /// Order-1.
     One,
 }