+++ noodles-cram/CHANGELOG.md
@@ -12,6 +12,8 @@
 
   * cram/codecs: Add fqzcomp encoder.
 
+  * cram/codecs: Increase visibility of `rans_4x8::Order`.
+
   * cram/codecs/name_tokenizer/encode: Add `DIGITS0`, `DIGITS`, `DELTA`,
     `DELTA0`, and `MATCH` tokens.